git: add icons
Also break services out into individual files
